Update file struct and README.md
